+++ packages/plugin-oauth-wechat/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swr-login/plugin-oauth-wechat
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/core@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes