+++ packages/plugin-passkey/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swr-login/plugin-passkey
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/core@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes