+++ packages/swr-login/CHANGELOG.md
@@ -1,5 +1,13 @@
 # swr-login
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/core@0.3.0
+  - @swr-login/react@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes